+++ src/CompView.cpp
@@ -23,12 +23,11 @@ void CompView::moveUnits(float deltaS){
 
   // Poll the actor class for its speed and position on map, use the shortest distance >= 0
   //	and move that direction
-
+  /*
   vector<vector<int>> dists=gameLogic->getDistances();
   vector<vector<int>> aboveFloor=gameLogic->getAboveFloor();
 
   for(auto iterator : waveManager->getSpawnedEnemyUnits()){
-
     shared_ptr<MeleeUnit> currentUnit= iterator.second;
 
     if(!coordsInsideTile(currentUnit->getRow(),currentUnit->getCol(),currentUnit->getXCoordinate(),currentUnit->getXCoordinate())){
@@ -71,6 +70,7 @@ void CompView::moveUnits(float deltaS){
       currentUnit->update(deltaS);
     }
   }
+  */
 }