+++ src/Projectiles/CheesePizzaProjectile.cpp
@@ -10,5 +10,11 @@ CheesePizzaProjectile::CheesePizzaProjectile(shared_ptr<TextLoader> textLoader, 
   this -> radius = textLoader->getInteger(string("IDS_CP_AR"));;
   this -> actorTypeID = textLoader->getTypeID(string("IDS_CPP"));
   this -> textures = textureLoader -> getTexture(actorTypeID);
-  this -> initSprite();
+  //set the initial sprite texture
+  this ->current_sprite = 0;
+  //load in the initial texture for sizing
+  initSprite();
+
+  //set the sprite for the actor to have a position that is equivalent to its center
+  setToCenter();
 }