+++ src/Projectiles/MiniMMSProjectile.cpp
@@ -9,5 +9,11 @@ MiniMMSProjectile::MiniMMSProjectile(shared_ptr<TextLoader> textLoader, shared_p
   this -> radius = textLoader->getInteger(string("IDS_MMM_AR"));;
   this -> actorTypeID = textLoader->getTypeID(string("IDS_MMMP"));
   this -> textures = textureLoader -> getTexture(actorTypeID);
-  this -> initSprite();
+  //set the initial sprite texture
+  this ->current_sprite = 0;
+  //load in the initial texture for sizing
+  initSprite();
+
+  //set the sprite for the actor to have a position that is equivalent to its center
+  setToCenter();
 }