+++ src/Units/WaffleFryUnit.cpp
@@ -11,5 +11,11 @@ WaffleFryUnit::WaffleFryUnit(shared_ptr<TextLoader> textLoader, shared_ptr<Event
   this->lunchMoney=textLoader->getInteger(string("IDS_WF_LM"));;
   this->actorTypeID=textLoader->getTypeID(string("IDS_WFU"));
   this -> textures = textureLoader -> getTexture(actorTypeID);
-  this -> initSprite();
+  //set the initial sprite texture
+  this ->current_sprite = 0;
+  //load in the initial texture for sizing
+  initSprite();
+
+  //set the sprite for the actor to have a position that is equivalent to its center
+  setToCenter();
 }